Extract SampleBatchInformationSpreadsheet (#435)
* Extract component

* remove private delegating method

* Extend sample info by sample code

* Fix smaller issues

* add sample id to sampleInfo

* make sure the column can be locked


* Only add sample code if requested

* Fix typo

---------
